Small bug fix,README updates, and improved error messages
The bug fix had to do with Go being pretty particular about
the ordering of 'ResponseWriter.Write(()' methods resulting
in the 'Location' header not being written. The README now
has more information about bulk updates and a bit more info
on the API. The error messages are somewhat improved as well.